refactor: Point 리팩토링

+++ apps/commerce-api/src/main/java/com/loopers/domain/point/PointInfo.java
@@ -1,6 +1,6 @@
 package com.loopers.domain.point;
 
-public record PointInfo(int balance) {
+public record PointInfo(long balance) {
     public static PointInfo from(PointEntity entity) {
         return new PointInfo(
                 entity.getBalance()
